backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
pyyaml
openai
tiktoken
qdrant-client
requests
python-on-whales